backend/requirements.txt
# Runtime dependencies for the prediction pipeline.
# Dev / tooling dependencies live in ../requirements-dev.txt.

# Data
pandas>=2.0,<3
numpy>=1.24,<3
pyarrow>=14.0,<22

# ML
scikit-learn>=1.4,<2
xgboost>=2.0,<4
lightgbm>=4.0,<5
scipy>=1.11,<2

# HTTP / scraping / ingestion
requests>=2.31,<3
beautifulsoup4>=4.12,<5
lxml>=4.9,<6
statsbombpy>=1.1,<2

# Config + utilities
pyyaml>=6.0,<7
python-dotenv>=1.0,<2
tqdm>=4.66,<5
loguru>=0.7,<1
pydantic>=2.0,<3
